+++ spark-3.4/clickhouse-spark/src/main/scala/com/clickhouse/spark/ClickHouseTable.scala
@@ -161,6 +161,7 @@ case class ClickHouseTable(
   }
 
   override def newWriteBuilder(info: LogicalWriteInfo): ClickHouseWriteBuilder = {
+    val writeOptions = new WriteOptions(info.options.asCaseSensitiveMap())
     val writeJob = WriteJobDescription(
       queryId = info.queryId,
       tableSchema = schema,
@@ -176,7 +177,8 @@ case class ClickHouseTable(
       shardingKey = shardingKey,
       partitionKey = partitionKey,
       sortingKey = sortingKey,
-      writeOptions = new WriteOptions(info.options.asCaseSensitiveMap()),
+      writeOptions = writeOptions,
+      writeSettings = writeOptions.settings,
       functionRegistry = functionRegistry
     )