+++ spark-4.0/clickhouse-spark/src/main/scala/com/clickhouse/spark/ClickHouseTable.scala
@@ -171,6 +171,7 @@ case class ClickHouseTable(
   }
 
   override def newWriteBuilder(info: LogicalWriteInfo): ClickHouseWriteBuilder = {
+    val writeOptions = new WriteOptions(info.options.asCaseSensitiveMap())
     val writeJob = write.WriteJobDescription(
       queryId = info.queryId,
       tableSchema = schema,
@@ -186,7 +187,8 @@ case class ClickHouseTable(
       shardingKey = shardingKey,
       partitionKey = partitionKey,
       sortingKey = sortingKey,
-      writeOptions = new WriteOptions(info.options.asCaseSensitiveMap()),
+      writeOptions = writeOptions,
+      writeSettings = writeOptions.settings,
       functionRegistry = functionRegistry
     )